+++ forge/src/main/java/dev/mayaqq/estrogen/forge/EstrogenForge.java
@@ -4,10 +4,14 @@ import com.simibubi.create.foundation.config.ConfigBase;
 import dev.mayaqq.estrogen.Estrogen;
 import dev.mayaqq.estrogen.client.registry.EstrogenRenderer;
 import dev.mayaqq.estrogen.config.EstrogenConfig;
+import dev.mayaqq.estrogen.registry.EstrogenPotatoProjectiles;
 import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLEnvironment;
 
 import java.util.Map;
@@ -22,11 +26,18 @@ public class EstrogenForge {
         for (Map.Entry<ModConfig.Type, ConfigBase> pair : EstrogenConfig.CONFIGS.entrySet())
             ModLoadingContext.get().registerConfig(pair.getKey(), pair.getValue().specification);
 
+        IEventBus modEventBus = FMLJavaModLoadingContext.get().getModEventBus();
+
         // Init Estrogen main class
         Estrogen.init();
 
         if(FMLEnvironment.dist == Dist.CLIENT) {
             EstrogenRenderer.register();
         }
+        modEventBus.addListener(EstrogenForge::init);
+    }
+
+    public static void init(final FMLCommonSetupEvent event) {
+        event.enqueueWork(EstrogenPotatoProjectiles::register);
     }
 }