+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-# Estrogen `4.2.2` `1.20.1` `Fabric & Forge`
+# Estrogen `@VERSION@` `1.20.1` `Fabric & Forge`
 ### Changelog:
-- Fixed issue with mob drops on Fabric
+- Updated music
+- Fixed crash on forge